feat(cache): add caching for TMDb API requests
