+++ package.json
@@ -36,6 +36,7 @@
     "intl": "^1.2.5",
     "lodash": "^4.17.20",
     "next": "10.0.3",
+    "node-cache": "^5.1.2",
     "node-schedule": "^1.3.3",
     "nodemailer": "^6.4.17",
     "nookies": "^2.5.2",
