+++ server/api/themoviedb/constants.ts
@@ -1,0 +1,1 @@
+export const ANIME_KEYWORD_ID = 210024;
