+++ server/entity/MediaRequest.ts
@@ -15,7 +15,8 @@ import { User } from './User';
 import Media from './Media';
 import { MediaStatus, MediaRequestStatus, MediaType } from '../constants/media';
 import { getSettings } from '../lib/settings';
-import TheMovieDb, { ANIME_KEYWORD_ID } from '../api/themoviedb';
+import TheMovieDb from '../api/themoviedb';
+import { ANIME_KEYWORD_ID } from '../api/themoviedb/constants';
 import RadarrAPI from '../api/radarr';
 import logger from '../logger';
 import SeasonRequest from './SeasonRequest';
