+++ server/job/sonarrsync/index.ts
@@ -2,7 +2,8 @@ import { uniqWith } from 'lodash';
 import { getRepository } from 'typeorm';
 import { v4 as uuid } from 'uuid';
 import SonarrAPI, { SonarrSeries } from '../../api/sonarr';
-import TheMovieDb, { TmdbTvDetails } from '../../api/themoviedb';
+import TheMovieDb from '../../api/themoviedb';
+import { TmdbTvDetails } from '../../api/themoviedb/interfaces';
 import { MediaStatus, MediaType } from '../../constants/media';
 import Media from '../../entity/Media';
 import Season from '../../entity/Season';
